chore: move doctor to engine-server
