+++ packages/dendron-cli/src/commands/backfillV2.ts
@@ -1,10 +1,5 @@
-import {
-  DEngineClient,
-  genUUID,
-  NoteProps,
-  NotePropsDict,
-  NoteUtils,
-} from "@dendronhq/common-all";
+import { DEngineClient, NoteProps } from "@dendronhq/common-all";
+import { BackfillService } from "@dendronhq/engine-server";
 import _ from "lodash";
 import { BaseCommand, CommandCommonProps } from "./base";
 
@@ -22,33 +17,7 @@ type CommandOutput = CommandCommonProps;
 
 export class BackfillV2Command extends BaseCommand<CommandOpts, CommandOutput> {
   async execute(opts: CommandOpts): Promise<CommandCommonProps> {
-    const { engine, note, overwriteFields } = _.defaults(opts, {
-      overwriteFields: [],
-    });
-    const candidates: NotePropsDict = _.isUndefined(note)
-      ? engine.notes
-      : { [note.id]: note };
-    const notes = await Promise.all(
-      _.values(candidates)
-        .filter((n) => !n.stub)
-        .map(async (n) => {
-          overwriteFields.forEach((f) => {
-            if (f === "title") {
-              n.title = NoteUtils.genTitle(n.fname);
-            } else if (f === "id") {
-              // ids starting or ending with `-` or `_` break pages in Github publishing
-              if (n.id.match(/^[-_]|[-_]$/)) {
-                n.id = genUUID();
-              }
-            } else {
-              throw Error(`unknown overwrite field: ${f}`);
-            }
-          });
-          return n;
-        })
-    );
-    // @ts-ignore
-    await engine.store.bulkAddNotes({ notes });
-    return {};
+    const backfillService = new BackfillService();
+    return backfillService.updateNotes(opts);
   }
 }
