+++ packages/engine-server/package.json
@@ -40,6 +40,7 @@
     "@dendronhq/common-all": "^0.70.0",
     "@dendronhq/common-server": "^0.70.0",
     "@dendronhq/remark-mermaid": "^0.2.0",
+    "@jcoreio/async-throttle": "^1.4.3",
     "@mapbox/rehype-prism": "^0.5.0",
     "@types/fs-extra": "^9.0.1",
     "@types/klaw": "^3.0.1",
