+++ packages/engine-server/src/index.ts
@@ -20,3 +20,5 @@ export * from "./metadata";
 export { execa };
 export * from "./util/inMemoryNoteCache";
 export * from "./drivers";
+export * from "./doctor";
+export * from "./backfillV2";
