+++ packages/engine-test-utils/src/__tests__/dendron-cli/commands/doctor.spec.ts
@@ -1,11 +1,8 @@
 import { WorkspaceOpts } from "@dendronhq/common-all";
 import { file2Note } from "@dendronhq/common-server";
+import { DoctorActions } from "@dendronhq/engine-server";
 import { AssertUtils, NoteTestUtilsV4 } from "@dendronhq/common-test-utils";
-import {
-  DoctorActions,
-  DoctorCLICommand,
-  DoctorCLICommandOpts,
-} from "@dendronhq/dendron-cli";
+import { DoctorCLICommand, DoctorCLICommandOpts } from "@dendronhq/dendron-cli";
 import path from "path";
 import fs from "fs-extra";
 import { createEngineFromServer, runEngineTestV5 } from "../../../engine";
