+++ packages/plugin-core/src/test/suite-integ/DoctorCommand.test.ts
@@ -1,7 +1,7 @@
 import { NoteProps, NoteUtils, VaultUtils } from "@dendronhq/common-all";
 import { vault2Path } from "@dendronhq/common-server";
 import { NoteTestUtilsV4 } from "@dendronhq/common-test-utils";
-import { DoctorActions } from "@dendronhq/dendron-cli";
+import { DoctorActions } from "@dendronhq/engine-server";
 import { ENGINE_HOOKS } from "@dendronhq/engine-test-utils";
 import fs from "fs-extra";
 import _ from "lodash";
